Update and rename homework/Kupreichik Anton/FinalWork/src/main/java/com.company/service/Category.java to homework/Kupreichik Anton/FinalWork/src/main/java/com.company/bean/Category.java
